correct answer to day 2 puzzle 1

+++ 2025/02/ConsoleApp1/ConsoleApp1/Program.cs
@@ -1,7 +1,7 @@
 ﻿// See https://aka.ms/new-console-template for more information
 using System.IO;
 
-int invalidTotal = 0;
+long invalidTotal = 0;
 string content = File.ReadAllText("../../../input.txt");
 Console.WriteLine(content);
 Console.WriteLine("Hello, World!");
@@ -9,8 +9,8 @@ string[] ranges = content.Split(",");
 foreach (string range in ranges)
 {
     string[] keys = range.Split("-");
-    int key1 = Int32.Parse(keys[0]);
-    int key2 = Int32.Parse(keys[1]);
+    long key1 = Int64.Parse(keys[0]);
+    long key2 = Int64.Parse(keys[1]);
     // Console.WriteLine(key1 + " != " + key2);
     for (int i = 0; i < key2 - key1 + 1; i++)
     {
